// src/sub-module/authentication/model/register.model.ts
export class Register {
  createBy: any = null;
  updateBy: any = null;
  deleteby: any = null;
  userId: any = null;
  username: any = null;
  password: any = null;
  provider: any = null;
  phoneNo: any = null;
  dob: any = null;
  profileImage: any = null;
  firstName: any = null;
  lastName: any = null;
  email: any = null;
}